requirements.txt
fastapi
uvicorn[standard]
python-dotenv
SQLAlchemy
mysqlclient
aiomysql
PyMySQL
motor
pydantic-settings
cryptography
aiofiles